+++ src/greentest/test__queue.py
@@ -1,12 +1,14 @@
 import greentest
-from greentest import TestCase, main, GenericGetTestCase
+from greentest import TestCase, main
 import gevent
 from gevent.hub import get_hub, LoopExit
 from gevent import util
 from gevent import queue
 from gevent.queue import Empty, Full
 from gevent.event import AsyncResult
+from greentest.timing import AbstractGenericGetTestCase
 
+# pylint:disable=too-many-ancestors
 
 class TestQueue(TestCase):
 
@@ -372,7 +374,7 @@ class TestJoinEmpty(TestCase):
         q.join()
 
 
-class TestGetInterrupt(GenericGetTestCase):
+class TestGetInterrupt(AbstractGenericGetTestCase):
 
     Timeout = Empty
 
@@ -397,7 +399,7 @@ class TestGetInterruptChannel(TestGetInterrupt):
     kind = queue.Channel
 
 
-class TestPutInterrupt(GenericGetTestCase):
+class TestPutInterrupt(AbstractGenericGetTestCase):
     kind = queue.Queue
     Timeout = Full
 
@@ -430,7 +432,7 @@ class TestPutInterruptChannel(TestPutInterrupt):
         return self.kind()
 
 
-del GenericGetTestCase
+del AbstractGenericGetTestCase
 
 
 if __name__ == '__main__':